requirements.txt
scikit-learn>=1.3.0,<2.0
xgboost>=2.0.0,<3.0
shap>=0.43.0,<1.0
pandas>=2.1.0,<3.0
numpy>=1.25.0,<2.0
fastapi>=0.104.0,<1.0
uvicorn>=0.24.0,<1.0
pydantic>=2.5.0,<3.0
mlflow>=2.9.0,<3.0
joblib>=1.3.0,<2.0
pyyaml>=6.0,<7.0
httpx>=0.25.0,<1.0
pytest>=7.4.0,<9.0
